+++ app/models/common_data_model/cdm_demographic.rb
@@ -1,5 +1,0 @@
-class CdmDemographic < ActiveRecord::Base
-  self.table_name = 'cdm_demographic'
-
-  belongs_to :user, foreign_key: 'patid'
-end