+++ app/models/common_data_model/cdm_enrollment.rb
@@ -1,6 +1,0 @@
-class CdmEnrollment < ActiveRecord::Base
-  self.table_name = 'cdm_enrollment'
-
-  belongs_to :user, foreign_key: 'patid'
-
-end