+++ app/models/common_data_model/cdm_pro_cm.rb
@@ -1,6 +1,0 @@
-class CdmProCm < ActiveRecord::Base
-  self.table_name = 'cdm_pro_cm'
-
-  belongs_to :user, foreign_key: 'patid'
-
-end