+++ app/models/common_data_model/cdm_vital.rb
@@ -1,6 +1,0 @@
-class CdmVital < ActiveRecord::Base
-  self.table_name = 'cdm_vital'
-
-  belongs_to :user, foreign_key: 'patid'
-
-end